+++ app/pages/legal.js
@@ -1,14 +1,10 @@
 const html = require('choo/html');
 const raw = require('choo/html/raw');
-const assets = require('../../common/assets');
 const title = require('../templates/title');
 
 module.exports = function(state) {
   return html`
-    <div>
-      <a href="/" class="goBackButton"> 
-        <img src="${assets.get('back-arrow.svg')}"/> 
-      </a>
+    <div class="page">
       ${title(state)}
       <div class="title">${state.translate('legalHeader')}</div>
         ${raw(
