+++ app/templates/fileIcon/fileIcon.css
@@ -4,7 +4,7 @@
   pointer-events: none;
   margin: 8px;
   color: #fff;
-  background-image: url('../assets/red_file.svg');
+  background-image: url('../assets/blue_file.svg');
   width: 22px;
   height: 32px;
   overflow: hidden;
