+++ app/templates/fxPromo/fxPromo.css
@@ -10,10 +10,6 @@
   align-content: center;
   align-items: center;
   justify-content: center;
-  position: fixed;
-  top: 0;
-  right: 0;
-  left: 0;
 }
 
 .fxPromo > div {
