+++ app/templates/header/header.css
@@ -1,14 +1,25 @@
 .header {
-  align-items: flex-end;
+  flex: none;
   display: flex;
   flex-direction: row;
-  height: calc(var(--grid-basis) * 4);
+  align-items: center;
   justify-content: space-between;
   width: 100%;
+  height: 64px;
+  background-color: white;
+  box-shadow: 0 -1px 0 0 #d7d7db7f, 0 2px 6px 0 #0000001e;
 }
 
 .header h1 {
   margin: 0;
+}
+
+.header__controls {
+  justify-self: end;
+  display: grid;
+  grid-auto-flow: column;
+  grid-gap: 8px;
+  align-items: center;
 }
 
 .feedback {
@@ -26,6 +37,4 @@
   text-indent: 2px;
   transition: all 250ms ease-in-out;
   white-space: nowrap;
-  margin-left: calc(var(--grid-basis) * 2);
-  margin-bottom: 8px;
 }
