+++ app/templates/setPasswordSection/setPasswordSection.css
@@ -5,7 +5,6 @@
 }
 
 .checkbox {
-  flex: auto;
   height: 24px;
 }
