+++ app/templates/uploadedFileList/uploadedFileList.css
@@ -1,7 +1,5 @@
 .uploadedFiles {
-  border: 1px solid rgba(12, 12, 13, 0.1);
-  border-radius: 4px;
-  box-sizing: border-box;
+  background: rgba(132, 193, 255, 0.1);
   margin: 0;
   padding: 0;
   align-content: center;
